backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
openai==0.28.0
python-dotenv==1.0.1
PyMuPDF==1.24.9
scikit-learn==1.5.1
numpy==2.0.1
python-multipart==0.0.9
